share close

+++ mywords-go/model/mtype/share_info.go
@@ -1,6 +1,6 @@
 package mtype
 
 type ShareInfo struct {
-	Port int   `json:"port"`
+	Port int64 `json:"port"`
 	Code int64 `json:"code"`
 }
